Drop usage of pixel x/y in npcs

+++ server/include/TLevel.h
@@ -218,9 +218,9 @@ class TLevel
 		CString getChestStr(const TLevelChest& chest) const;
 
 #ifdef V8NPCSERVER
-		std::vector<TNPC *> findAreaNpcs(int pX, int pY, int pWidth, int pHeight);
-		std::vector<TNPC*> testTouch(int pX, int pY);
-		TNPC *isOnNPC(int pX, int pY, bool checkEventFlag = false);
+		std::vector<TNPC *> findAreaNpcs(float pX, float pY, int pWidth, int pHeight);
+		std::vector<TNPC*> testTouch(float pX, float pY);
+		TNPC *isOnNPC(float pX, float pY, bool checkEventFlag = false);
 		void sendChatToLevel(const TPlayer *player, const std::string& message);
 
 		IScriptObject<TLevel>* getScriptObject() const;
